requirements.txt
selenium==4.15.2
pandas==2.1.3
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
